requirements/development.txt
-r base.txt
django-debug-toolbar==4.4.6
django-extensions==3.2.3
ipython==8.27.0
factory-boy==3.3.1
coverage==7.6.1
pytest-django==4.9.0
